// imxweb/projects/qbm/src/lib/appConfig/appconfig.interface.ts
/*
 * ONE IDENTITY LLC. PROPRIETARY INFORMATION
 *
 * This software is confidential.  One Identity, LLC. or one of its affiliates or
 * subsidiaries, has supplied this software to you under terms of a
 * license agreement, nondisclosure agreement or both.
 *
 * You may not copy, disclose, or use this software except in accordance with
 * those terms.
 *
 *
 * Copyright 2025 One Identity LLC.
 * ALL RIGHTS RESERVED.
 *
 * ONE IDENTITY LLC. MAKES NO REPRESENTATIONS OR
 * WARRANTIES ABOUT THE SUITABILITY OF THE SOFTWARE,
 * EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED
 * TO THE IMPLIED WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE, OR
 * NON-INFRINGEMENT.  ONE IDENTITY LLC. SHALL NOT BE
 * LIABLE FOR ANY DAMAGES SUFFERED BY LICENSEE
 * AS A RESULT OF USING, MODIFYING OR DISTRIBUTING
 * THIS SOFTWARE OR ITS DERIVATIVES.
 *
 */

import { TranslationConfiguration } from './translationConfiguration.interface';
import { RouteConfig } from './route-config.interface';

export interface AppConfig {
  Basepath?: string;
  Title: string;
  WebAppIndex: string;
  WebAppIdentifier?: string;
  NotificationUpdateInterval?: number;
  DatastoreIssueTreshold?: number;
  Translation?: TranslationConfiguration;
  routeConfig?: RouteConfig;

  /** Do not show the app name when showing the product name, in case
   * the app name is redundant.
   */
  DoNotShowAppNameWithProduct?: boolean;
  LocalDocPath?: { [translationKey: string]: string };
}